components/job_execution/python/requirements.txt
PyYAML==5.2
python-nomad==1.2.1
